Add generate-responses-and-expected-results-for-smartdown-flow.rb
This is mostly copied from
generate-responses-and-expected-results-for-smart-answer.rb, but has been
updated with Smartdown specific code.

This appears to be generating the YAML file correctly!

TODO: Avoid hardcoding the name of the flow.

TODO: Work out why the generated rendered outcomes don't get past the "Why type
of student are you" question.