Move formatting of total_hourly_rate to templates
Formatting is a view concern rather than a flow concern.

I've moved the formatting of `total_hourly_rate` from the flow to the ERB
templates for am-i-getting-minimum-wage and minimum-wage-calculator-employers.

The `MinimumWageCalculator#format_money` and `OutcomeHelper#format_money`
methods both remove the zeros for whole numbers. The
`OutcomeHelper#format_money` also adds separators to large numbers and prepends
the currency symbol. The `OutcomeHelper#format_money` method also adds the
minus sign ('-') before the pound symbol ('£') rather than after it. So,
negative numbers now appear as, for example, -£55.92 instead of £-55.92.

I've updated a number of the test artefacts that have changed now that larger
numbers include a separator, and the minus sign has moved.