Remove `total_hourly_rate` precalculate block
Using `MinimumWageCalculator#total_hourly_rate` in the ERB template allows us
to reduce the number of methods that we have to make available to the view.

I've chosen to update some of the tests that were previously asserting state
variables so that they now assert against the `MinimumWageCalculator`. I'm not
convinced about this but I think it means that I'm not accidentally losing
potentially important/interesting tests.